added backtracking problems
