+++ src/test/java/org/milan/datastructure/ExponentialSearchTest.java
@@ -2,7 +2,7 @@ package org.milan.datastructure;
 
 import org.junit.Assert;
 import org.junit.Test;
-import org.milan.datastructure.searching.ExponentialSearch;
+import org.milan.algorithm.searching.ExponentialSearch;
 
 /**
  * Test Class for {@link ExponentialSearch}
