--- src/test/java/org/milan/datastructure/bits/AddNumberTest.java
@@ -1,20 +1,0 @@
-package org.milan.datastructure.bits;
-
-import org.junit.Assert;
-import org.junit.Test;
-
-/**
- * Test Class for {@link AddNumber}
- *
- * @author Milan Rathod
- */
-public class AddNumberTest {
-
-    @Test
-    public void testAdd() {
-        AddNumber addNumber = new AddNumber();
-        int result = addNumber.add(11);
-
-        Assert.assertEquals(12, result);
-    }
-}
